Extract hash tests to separate file
