+++ .github/workflows/test.yml
@@ -28,24 +28,8 @@ jobs:
       matrix:
         redis-version: [ "6.2.6", "7.0.5" ]
         python-version: [ "3.7", "3.8", "3.10", '3.11' ]
-        redis-py: [ "4.1.2", "4.3.4" ]
+        redis-py: [ '4.3.4' ]
         include:
-          - python-version: "3.10"
-            redis-version: "6.2.6"
-            redis-py: "2.10.6"
-            aioredis: "1.3.1"
-          - python-version: "3.10"
-            redis-version: "6.2.6"
-            redis-py: "3.5.3"
-            aioredis: "1.3.1"
-          - python-version: "3.10"
-            redis-version: "6.2.6"
-            redis-py: "4.0.1"
-            aioredis: "1.3.1"
-          - python-version: "3.10"
-            redis-version: "6.2.6"
-            redis-py: "4.1.2"
-            aioredis: "2.0.1"
           - python-version: "3.10" # should work fine with redis.asyncio
             redis-version: "7.0.5"
             redis-py: "4.3.4"
