+++ test/test_fakeredis6.py
@@ -9,14 +9,11 @@ from time import sleep
 import pytest
 import redis
 import redis.client
-from packaging.version import Version
 from redis.exceptions import ResponseError
 
 import fakeredis
 import testtools
 from testtools import raw_command
-
-REDIS_VERSION = Version(redis.__version__)
 
 
 def key_val_dict(size=100):
@@ -30,10 +27,7 @@ def round_str(x):
 
 
 def zincrby(r, key, amount, value):
-    if REDIS_VERSION >= Version('3'):
-        return r.zincrby(key, amount, value)
-    else:
-        return r.zincrby(key, value, amount)
+    return r.zincrby(key, amount, value)
 
 
 def test_large_command(r):
@@ -491,11 +485,7 @@ def test_mget(r):
 
 
 def test_mget_with_no_keys(r):
-    if REDIS_VERSION >= Version('3'):
-        assert r.mget([]) == []
-    else:
-        with pytest.raises(redis.ResponseError, match='wrong number of arguments'):
-            r.mget([])
+    assert r.mget([]) == []
 
 
 def test_mget_mixed_types(r):
@@ -578,7 +568,7 @@ def test_set_px_using_timedelta(r):
     assert r.get('foo') == b'bar'
 
 
-@pytest.mark.skipif(REDIS_VERSION < Version('3.5'), reason="Test is only applicable to redis-py 3.5+")
+@testtools.run_test_if_redispy_ver('below', '3.5')
 @pytest.mark.min_server('6.0')
 def test_set_keepttl(r):
     r.set('foo', 'bar', ex=100)
@@ -592,7 +582,7 @@ def test_set_conflicting_expire_options(r):
         r.set('foo', 'bar', ex=1, px=1)
 
 
-@pytest.mark.skipif(REDIS_VERSION < Version('3.5'), reason="Test is only applicable to redis-py 3.5+")
+@testtools.run_test_if_redispy_ver('below', '3.5')
 def test_set_conflicting_expire_options_w_keepttl(r):
     with pytest.raises(ResponseError):
         r.set('foo', 'bar', ex=1, keepttl=True)
@@ -754,246 +744,6 @@ def test_delete_nonexistent_key(r):
     assert r.delete('foo') == 0
 
 
-
-# Tests for the hash type.
-
-def test_hstrlen_missing(r):
-    assert r.hstrlen('foo', 'doesnotexist') == 0
-
-    r.hset('foo', 'key', 'value')
-    assert r.hstrlen('foo', 'doesnotexist') == 0
-
-
-def test_hstrlen(r):
-    r.hset('foo', 'key', 'value')
-    assert r.hstrlen('foo', 'key') == 5
-
-
-def test_hset_then_hget(r):
-    assert r.hset('foo', 'key', 'value') == 1
-    assert r.hget('foo', 'key') == b'value'
-
-
-def test_hset_update(r):
-    assert r.hset('foo', 'key', 'value') == 1
-    assert r.hset('foo', 'key', 'value') == 0
-
-
-def test_hset_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hset('foo', 'key', 'value')
-
-
-def test_hgetall(r):
-    assert r.hset('foo', 'k1', 'v1') == 1
-    assert r.hset('foo', 'k2', 'v2') == 1
-    assert r.hset('foo', 'k3', 'v3') == 1
-    assert r.hgetall('foo') == {
-        b'k1': b'v1',
-        b'k2': b'v2',
-        b'k3': b'v3'
-    }
-
-
-def test_hgetall_empty_key(r):
-    assert r.hgetall('foo') == {}
-
-
-def test_hgetall_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hgetall('foo')
-
-
-def test_hexists(r):
-    r.hset('foo', 'bar', 'v1')
-    assert r.hexists('foo', 'bar') == 1
-    assert r.hexists('foo', 'baz') == 0
-    assert r.hexists('bar', 'bar') == 0
-
-
-def test_hexists_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hexists('foo', 'key')
-
-
-def test_hkeys(r):
-    r.hset('foo', 'k1', 'v1')
-    r.hset('foo', 'k2', 'v2')
-    assert set(r.hkeys('foo')) == {b'k1', b'k2'}
-    assert set(r.hkeys('bar')) == set()
-
-
-def test_hkeys_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hkeys('foo')
-
-
-def test_hlen(r):
-    r.hset('foo', 'k1', 'v1')
-    r.hset('foo', 'k2', 'v2')
-    assert r.hlen('foo') == 2
-
-
-def test_hlen_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hlen('foo')
-
-
-def test_hvals(r):
-    r.hset('foo', 'k1', 'v1')
-    r.hset('foo', 'k2', 'v2')
-    assert set(r.hvals('foo')) == {b'v1', b'v2'}
-    assert set(r.hvals('bar')) == set()
-
-
-def test_hvals_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hvals('foo')
-
-
-def test_hmget(r):
-    r.hset('foo', 'k1', 'v1')
-    r.hset('foo', 'k2', 'v2')
-    r.hset('foo', 'k3', 'v3')
-    # Normal case.
-    assert r.hmget('foo', ['k1', 'k3']) == [b'v1', b'v3']
-    assert r.hmget('foo', 'k1', 'k3') == [b'v1', b'v3']
-    # Key does not exist.
-    assert r.hmget('bar', ['k1', 'k3']) == [None, None]
-    assert r.hmget('bar', 'k1', 'k3') == [None, None]
-    # Some keys in the hash do not exist.
-    assert r.hmget('foo', ['k1', 'k500']) == [b'v1', None]
-    assert r.hmget('foo', 'k1', 'k500') == [b'v1', None]
-
-
-def test_hmget_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hmget('foo', 'key1', 'key2')
-
-
-def test_hdel(r):
-    r.hset('foo', 'k1', 'v1')
-    r.hset('foo', 'k2', 'v2')
-    r.hset('foo', 'k3', 'v3')
-    assert r.hget('foo', 'k1') == b'v1'
-    assert r.hdel('foo', 'k1') == 1
-    assert r.hget('foo', 'k1') is None
-    assert r.hdel('foo', 'k1') == 0
-    # Since redis>=2.7.6 returns number of deleted items.
-    assert r.hdel('foo', 'k2', 'k3') == 2
-    assert r.hget('foo', 'k2') is None
-    assert r.hget('foo', 'k3') is None
-    assert r.hdel('foo', 'k2', 'k3') == 0
-
-
-def test_hdel_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hdel('foo', 'key')
-
-
-def test_hincrby(r):
-    r.hset('foo', 'counter', 0)
-    assert r.hincrby('foo', 'counter') == 1
-    assert r.hincrby('foo', 'counter') == 2
-    assert r.hincrby('foo', 'counter') == 3
-
-
-def test_hincrby_with_no_starting_value(r):
-    assert r.hincrby('foo', 'counter') == 1
-    assert r.hincrby('foo', 'counter') == 2
-    assert r.hincrby('foo', 'counter') == 3
-
-
-def test_hincrby_with_range_param(r):
-    assert r.hincrby('foo', 'counter', 2) == 2
-    assert r.hincrby('foo', 'counter', 2) == 4
-    assert r.hincrby('foo', 'counter', 2) == 6
-
-
-def test_hincrby_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hincrby('foo', 'key', 2)
-
-
-def test_hincrbyfloat(r):
-    r.hset('foo', 'counter', 0.0)
-    assert r.hincrbyfloat('foo', 'counter') == 1.0
-    assert r.hincrbyfloat('foo', 'counter') == 2.0
-    assert r.hincrbyfloat('foo', 'counter') == 3.0
-
-
-def test_hincrbyfloat_with_no_starting_value(r):
-    assert r.hincrbyfloat('foo', 'counter') == 1.0
-    assert r.hincrbyfloat('foo', 'counter') == 2.0
-    assert r.hincrbyfloat('foo', 'counter') == 3.0
-
-
-def test_hincrbyfloat_with_range_param(r):
-    assert r.hincrbyfloat('foo', 'counter', 0.1) == pytest.approx(0.1)
-    assert r.hincrbyfloat('foo', 'counter', 0.1) == pytest.approx(0.2)
-    assert r.hincrbyfloat('foo', 'counter', 0.1) == pytest.approx(0.3)
-
-
-def test_hincrbyfloat_on_non_float_value_raises_error(r):
-    r.hset('foo', 'counter', 'cat')
-    with pytest.raises(redis.ResponseError):
-        r.hincrbyfloat('foo', 'counter')
-
-
-def test_hincrbyfloat_with_non_float_amount_raises_error(r):
-    with pytest.raises(redis.ResponseError):
-        r.hincrbyfloat('foo', 'counter', 'cat')
-
-
-def test_hincrbyfloat_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hincrbyfloat('foo', 'key', 0.1)
-
-
-def test_hincrbyfloat_precision(r):
-    x = 1.23456789123456789
-    assert r.hincrbyfloat('foo', 'bar', x) == x
-    assert float(r.hget('foo', 'bar')) == x
-
-
-def test_hsetnx(r):
-    assert r.hsetnx('foo', 'newkey', 'v1') == 1
-    assert r.hsetnx('foo', 'newkey', 'v1') == 0
-    assert r.hget('foo', 'newkey') == b'v1'
-
-
-def test_hmset_empty_raises_error(r):
-    with pytest.raises(redis.DataError):
-        r.hmset('foo', {})
-
-
-def test_hmset(r):
-    r.hset('foo', 'k1', 'v1')
-    assert r.hmset('foo', {'k2': 'v2', 'k3': 'v3'}) is True
-
-
-def test_hmset_wrong_type(r):
-    testtools.zadd(r, 'foo', {'bar': 1})
-    with pytest.raises(redis.ResponseError):
-        r.hmset('foo', {'key': 'value'})
-
-
-def test_empty_hash(r):
-    r.hset('foo', 'bar', 'baz')
-    r.hdel('foo', 'bar')
-    assert not r.exists('foo')
-
-
 def test_sadd(r):
     assert r.sadd('foo', 'member1') == 1
     assert r.sadd('foo', 'member1') == 0
@@ -1044,7 +794,7 @@ def test_scan_iter_multiple_pages_with_match(r):
     assert actual == set(all_keys)
 
 
-@pytest.mark.skipif(REDIS_VERSION < Version('3.5'), reason="Test is only applicable to redis-py 3.5+")
+@testtools.run_test_if_redispy_ver('below', '3.5')
 @pytest.mark.min_server('6.0')
 def test_scan_iter_multiple_pages_with_type(r):
     all_keys = key_val_dict(size=100)
@@ -2450,11 +2200,8 @@ def test_pipeline_no_commands(r):
     p = r.pipeline()
     p.watch('foo')
     r.set('foo', '2')
-    if REDIS_VERSION >= Version('3.4'):
-        with pytest.raises(redis.WatchError):
-            p.execute()
-    else:
-        assert p.execute() == []
+    with pytest.raises(redis.WatchError):
+        p.execute()
 
 
 def test_pipeline_failed_transaction(r):
@@ -2873,10 +2620,7 @@ def test_pubsub_run_in_thread(r):
         1,
         pytest.param(
             None,
-            marks=pytest.mark.skipif(
-                Version("3.2") <= REDIS_VERSION < Version("3.3"),
-                reason="This test is not applicable to redis-py 3.2"
-            )
+            marks=testtools.run_test_if_redispy_ver('above', '3.2')
         )
     ]
 )
